Update batteryTray.py
cast level to int, otherwise no match for 10.0.png

+++ batteryTray.py
@@ -63,7 +63,7 @@ By Jamie Lentin
             return f.read()
 
         b_level = round(float(slurp(BATT_NOW)) / float(slurp(BATT_FULL)) * 100)
-        b_file = IMAGE_LOC + "." + str(b_level / 10) + ".png"
+        b_file = IMAGE_LOC + "." + str(int(b_level / 10)) + ".png"
         self.tray.set_tooltip(
             "%s: %d%%" %
             (rstrip(slurp(BATT_STATE)), b_level)
